+++ app/unit-tests/src/org/commcare/fragments/personalId/EmailHelperTest.kt
@@ -8,7 +8,7 @@ import org.commcare.CommCareTestApplication
 import org.commcare.android.database.connect.models.ConnectUserRecord
 import org.commcare.android.database.connect.models.PersonalIdSessionData
 import org.commcare.connect.database.ConnectUserDatabaseUtil
-import org.commcare.connect.network.ApiPersonalId
+import org.commcare.connect.network.connectId.ApiPersonalId
 import org.commcare.personalId.PersonalIdRecoveryCompleter
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
